Pairwise Correlation Data Frame

Returns a correlation data frame between all pairs of predictors in a training data frame. If the argument Non-numeric predictors are transformed into numeric via target encoding, using the 'response' variable as reference.

Attempts to handle correlations between pairs of variables of different types, as follows:

numeric vs. numeric: computed with stats::cor() with the methods "pearson" or "spearman" using cor_numerics().

numeric vs. character via cor_numerics_and_characters(), two alternatives leading to different results:

'response' is provided: the character variable is target-encoded as numeric using the values of the response as reference, and then its correlation with the numeric variable is computed with stats::cor(). This option generates a response-specific result suitable for training statistical and machine-learning models

'response' is NULL (or the name of a non-numeric column): the character variable is target-encoded as numeric using the values of the numeric predictor (instead of the response) as reference, and then their correlation is computed with stats::cor(). This option leads to a response-agnostic result suitable for clustering problems.

character vs. character, via cor_characters(), with two alternatives leading to different results:

'response' is provided: the character variables are target-encoded as numeric using the values of the response as reference, and then their correlation is computed with stats::cor().

response' is NULL (or the name of a non-numeric column): the association between the character variables is computed using Cramer's V. This option might be problematic, because R-squared values and Cramer's V, even when having the same range between 0 and 1, are not fully comparable.

Usage

cor_df(
  df = NULL,
  response = NULL,
  predictors = NULL,
  cor_method = "pearson",
  encoding_method = "mean"
)

cor_numerics(df = NULL, predictors = NULL, cor_method = "pearson")

cor_numerics_and_characters(
  df = NULL,
  predictors = NULL,
  cor_method = "pearson",
  encoding_method = "mean"
)

cor_characters(df, predictors)

Arguments

df

(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.

response

(optional, character string) Name of a numeric (non-numerics are ignored) response variable in df, only required if there are categorical variables named in the argument predictors. Default: NULL.

predictors

(optional; character vector) Names of the variables to select from df. If omitted, all numeric columns in df are used instead. If argument response is not provided, non-numeric variables are ignored. Default: NULL

cor_method

(optional; character string) Method used to compute pairwise correlations. Either "pearson" or "spearman". Default: "pearson".

encoding_method

(optional; character string). Name of the target encoding method to transform non-numeric variables to numeric. One of "mean", "rank", "loo", "rnorm" (see target_encoding_lab() for further details). Default: "mean"

Value

data frame; variable pairs and their correlation

data frame

data frame

data frame

See also

Author

Blas M. Benito, PhD

Examples

data(
  vi,
  vi_predictors
)

#reduce size of vi to speed-up example execution
vi <- vi[1:1000, ]
vi_predictors <- vi_predictors[1:10]

#without response
#categorical vs categorical compared with cramer_v()
#categorical vs numerical compared wit stats::cor() via target-encoding
#numerical vs numerical compared with stats::cor()
df <- cor_df(
  df = vi,
  predictors = vi_predictors
)

head(df)
#> # A tibble: 6 × 3
#>   x              y              correlation
#>   <chr>          <chr>                <dbl>
#> 1 swi_max        swi_mean             0.896
#> 2 swi_min        swi_mean             0.678
#> 3 swi_min        swi_max              0.642
#> 4 topo_diversity topo_slope           0.538
#> 5 topo_elevation topo_slope           0.379
#> 6 swi_mean       topo_diversity       0.243

#with response
#different solution than previous one
#because target encoding is done against the response
#rather than against the other numeric in the pair
df <- cor_df(
  df = vi,
  response = "vi_mean",
  predictors = vi_predictors
)

head(df)
#> # A tibble: 6 × 3
#>   x                  y                  correlation
#>   <chr>              <chr>                    <dbl>
#> 1 koppen_description koppen_zone              0.991
#> 2 koppen_group       koppen_zone              0.937
#> 3 koppen_description koppen_group             0.933
#> 4 swi_max            swi_mean                 0.896
#> 5 swi_mean           koppen_description       0.882
#> 6 swi_mean           koppen_zone              0.875