Plant richness and predictors of American ecoregions

Richness of vascular plants of the American ecoregions as defined in Ecoregions 2017, and computed from GBIF data, and environmental predictors.

Usage

data(ecoregions)

Format

A data frame with 225 rows and 49 columns:

ecoregion_id: Numeric id of the ecoregion.

ecoregion_name: Name of the ecoregion according to Ecoregions 2017.

x and y: Numeric, coordinates of the ecoregion centroids, in EPSG 4326.

plant_richness: Number of vascular species found in the ecoregion. Response variable.

plant_richness_binary: Binary version of plant_richness, with value 1 when plant_richness is higher than 5000, and 0 otherwise.

ecoregion_area_km2: Area of the ecoregion in squared kilometers.

sampling_bias: Indicator of sampling bias, measured as number of vascular plant species found in the ecoregion divided by the number of GBIF records of the ecoregion, divided by the area of the ecoregion.

neighbors_count, neighbors_area, neighbors_percent_shared_edge, and neighbors_average_aridity: Number, area sum, proportion of shared edge, and average aridity of the direct neighbors of each ecoregion.

climate_aridity_index_average: Average aridity index of the ecoregion.

climate_bio1_average to climate_bio15_average: Bioclimatic variables representing average annual temperature (BIO1), temperature_seasonality (BIO4), average and maximum temperature of the warmest month (BIO5), minimum, average and maximum monthly rainfall (BIO12), and precipitation seasonlality (BIO15).

climate_hypervolume: Volume of the climatic envelope of the ecoregion, computed with the hypervolume package.

landcover_bare_percent_average, landcover_herbs_percent_average, landcover_trees_percent_average, and landcover_ndvi_average: Average cover percentage of bare soil, herbs, and trees, and NDVI, extracted from MODIS Vegetation Continuous Fields and the MODIS NDVI/EVI product.

topography_elevation_average and topography_elevation_range: Average elevation of the ecoregion and its range.

fragmentation_ai to fragmentation_te: Geographic fragmentation indices of the ecoregion, computed with the R package landscapemetrics.

primary_productivity: Character. Ordered character version of the column landcover_ndvi_average.

landlocked: Binary. Defines whether a ecoregion is landlocked (value 1) or not (value 0).

dominant_landcover: Factor. Dominant landcover of the ecoregion (one of: "bare", "herbs", and "trees"), computed from the columns landcover_bare_percent_average, landcover_herbs_percent_average, and landcover_trees_percent_average.

See also