Association Between a Categorical Response and a Categorical Predictor

Computes Cramer's V, a measure of association between categorical or factor variables. Please see cor_cramer_v() for further details.

Usage

f_v(df)

Arguments

df

(required, data frame) with columns:

"x": (character or factor) categorical predictor.

"y": (character or factor) categorical response.

Value

numeric: Cramer's V

See also

Examples

#load example data
data(vi)

#reduce size to speed-up example
vi <- vi[1:1000, ]

#categorical response and predictor
#to data frame without NAs
df <- data.frame(
  y = vi[["vi_factor"]],
  x = vi[["soil_type"]]
) |>
  na.omit()

#Cramer's V
f_v(df = df)
#> [1] 0.5000691