Target encoding of character and factor variables

Target encoding involves replacing the values of categorical variables with continuous values that are indicative of the response variable. There are several approaches to target encoding, including calculating the mean target value for each level of the categorical variable. Target encoding can be useful for improving the performance of machine learning models, but it can also introduce bias if the encoded values are not representative of the true underlying relationship between the categorical variables and the target.

This function identifies categorical variables in the input data frame, transforms them using a set of methods defined by the user, and returns the input data frame with the newly encoded variables.

The target encoding methods implemented in this function are:

rank: returns the order of the group as a integer, being the 1 the rank of the group with the lower mean of of the response variable. This method accepts the noise argument, which adds white noise to the result to increase data variability and reduce leakage. The variables returned by this method are named with the suffix "__encoded_rank". This method is implemented in the function fe_target_encoding_rank().

mean: uses the mean value of the response over each group in the categorical variable. This option accepts noise. The variables returned by this method are named with the suffix "__encoded_mean". This method is implemented in the function fe_target_encoding_mean().

rnorm: This method computes the mean and standard deviation of the response for each group of the categorical variable, and uses rnorm() to generate values taken from a normal distribution. The argument sd.width is used as a multiplier of the standard deviation to reduce the range of values produced by rnorm() for each group of the categorical predictor. The variables returned by this method are named with the suffix "__encoded_rnorm". This method is implemented in the function fe_target_encoding_rnorm().

loo: This is the leave-one-out method. Each categorical value is replaced with the mean of the response variable across the other cases within the same group. The variables returned by this method are named with the suffix "__encoded_loo". This method is implemented in the function fe_target_encoding_loo().

The methods "mean" and "rank" support the noise argument. Values larger than zero in this argument add white noise to the target-encoded variables using stats::rnorm() via the function fe_target_encoding_noise()). The noise argument represents a fraction of the average differences between groups of the target-encoded variable. For example, if noise = 0.25 and the target-encoded variable has the unique values c(1, 2, 3), as it could be the case when using the "rank" method, then the average between-groups difference would be 1, and the range of the noise added to each row would go between 0 and 0.25

The method "rnorm" has the argument sd.width, which multiplies the standard deviation argument of the rnorm() function to limit the spread of the encoded values between groups.

Usage

fe_target_encoding(
  df = NULL,
  response = NULL,
  predictors = NULL,
  methods = c("mean", "rank", "loo", "rnorm"),
  sd.width = 0.1,
  seed = 1,
  noise = 0,
  replace = FALSE,
  verbose = TRUE
)

Arguments

df

(required; data frame, tibble, or sf) A training data frame. Default: NULL

response

(required; character string) Name of the response. Must be a column name of df. Default: NULL

predictors

(required; character vector) Names of all the predictors in df. Only character and factor predictors are processed, but all are returned in the "df" slot of the function's output. Default: NULL

methods

(optional; character string). Name of the target encoding methods. Default: c("mean", "rank", "loo", "rnorm")

sd.width

(optional; numeric vector) Only for the method "rnorm". Numeric vector with multiplicators of the standard deviation of each group in the categorical variable, in the range 0.01-1. Default: 0.1

seed

(optional; integer) Random seed to facilitate reproducibility. Default: 1

noise

(optional; numeric vector) Only in methods "mean" and "rank". Numeric vector with noise values in the range 0-1. Default: 0.

replace

(optional; logical) Advanced option that changes the behavior of the function. Use only if you really know exactly what you need. If TRUE, only the first option in the methods, noise, and sd.width arguments is used, it replaces each categorical variable with its encoded version, and returns the input data frame with the encoded variables and without the original ones. Default: FALSE

verbose

(optional; logical) If TRUE, messages and plots generated during the execution of the function are displayed. Default: TRUE

Value

If no target encoding is needed because all predictors are numeric, the function returns df.

Otherwise it returns a list of the class "fe_target_encoding" with the slots:

data: Input data frame, but with target-encoded character or factor columns.

leakage_test: Data frame with the results of a linear model between the target-encoded variable and the response aimed to identify potential data leakage. It contains the following columns:

encoded_predictor: name of the target-encoded variable.

correlation_with_response: R-squared resulting from cor.test() on the target-encoded variable and the response.

interpretation: Interpretation of the test, with the values "Leakage", "Likely leakage", "Unlikely leakage", and "No leakage". If you find concerning results, you may either increase the value of the noise argument (if method = "mean"), or select the "rnorm" method.

If the advanced option replace is TRUE, then the input data frame is returned with the categorical variables replaced with their encoded versions.

Examples

if(interactive()){

#loading example data
data(
  ecoregions,
  ecoregions_continuous_response,
  ecoregions_all_predictors
  )

#the dataframe ecoregions contains two categorical variables
unique(ecoregions$dominant_landcover)
unique(ecoregions$primary_productivity)

#applying all methods for a continuous response
output <- fe_target_encoding(
  data = ecoregions,
  response = ecoregions_continuous_response,
  predictors = ecoregions_all_predictors,
  methods = c(
    "mean",
    "rank",
    "rnorm",
    "loo"
  ),
  sd.width = c(0.01, 0.1, 1),
  noise = c(0, 1)
)

#the output has several objects
names(output)

#names of the encoded predictors
output$encoded_predictors

#the data with the original and the encoded predictors
colnames(output$data)

#a leakage test assessing the correlation between the response and the encoded predictors
output$leakage_test

#plotting the transformations of "primary_productivity"
tidyr::pivot_longer(
  data = output$data,
  cols = dplyr::all_of(
    grep(
      pattern = "primary_productivity",
      x = output$encoded_predictors,
      value = TRUE)
    )
) |>
  dplyr::select(
    plant_richness,
    primary_productivity,
    name,
    value
  ) |>
  ggplot2::ggplot() +
  ggplot2::aes(
    x = plant_richness,
    y = value,
    color = primary_productivity
  ) +
  ggplot2::facet_wrap(~name, scales = "free_y") +
  ggplot2::geom_point() +
  ggplot2::labs(
    x = "Response values",
    y = "Encoded values",
    color = "Original\ngroups"
  )

}