Compute the preference order for predictors based on a user-defined function.

Calculates the preference order of predictors based on a user-provided function that takes a predictor, a response, and a data frame as arguments.

Returns a data frame with the columns "predictor" and "preference". The former contains the predictors names in order, ready for the argument preference_order in cor_select(), vif_select() and collinear(). The latter contains the result of the function f for each combination of predictor and response.

Supports a parallelization setup via future::plan() and progress bars generated by the progressr package. See examples.

Usage

preference_order(
  df = NULL,
  response = NULL,
  predictors = NULL,
  f = f_rsquared,
  encoding_method = "mean"
)

Arguments

df

(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.

response

(optional, character string) Name of a numeric (non-numerics are ignored) response variable in df, only required if there are categorical variables named in the argument predictors. Default: NULL.

predictors

(optional; character vector) Names of the variables to select from df. If omitted, all numeric columns in df are used instead. If argument response is not provided, non-numeric variables are ignored. Default: NULL

f

(optional: function) Function returning a numeric value representing the relationship between a given predictor in predictors and response. The available options are:

f_rsquared() (default)

f_gam_deviance()

f_rf_rsquared()

f_logistic_auc_balanced()

f_logistic_auc_unbalanced()

f_gam_auc_balanced()

f_gam_auc_unbalanced()

f_rf_auc_balanced()

f_rf_auc_unbalanced()

encoding_method

(optional; character string). Name of the target encoding method to transform non-numeric variables to numeric. One of "mean", "rank", "loo", "rnorm" (see target_encoding_lab() for further details). Default: "mean"

Value

data frame

See also

Author

Blas M. Benito, PhD

Examples

data(
  vi,
  vi_predictors
)

#subset to limit example run time
vi <- vi[1:1000, ]
vi_predictors <- vi_predictors[1:20]

#parallelization setup
future::plan(
  future::multisession,
  workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
# progressr::handlers(global = TRUE)

#computing preference order
#with response
#numeric and categorical predictors in the output
#as the R-squared between each predictor and the response
preference.order <- preference_order(
  df = vi,
  response = "vi_mean",
  predictors = vi_predictors,
  f = f_rsquared
  )

preference.order
#>                 predictor  preference
#> 1             koppen_zone 0.813706530
#> 2      koppen_description 0.798401085
#> 3                 soil_ph 0.753180652
#> 4                swi_mean 0.743334183
#> 5            koppen_group 0.715080723
#> 6               soil_type 0.646557615
#> 7    soil_temperature_max 0.611299994
#> 8                 swi_max 0.593351577
#> 9  soil_temperature_range 0.541190402
#> 10              swi_range 0.418497231
#> 11                swi_min 0.260662568
#> 12               soil_soc 0.160172083
#> 13         topo_diversity 0.116888153
#> 14   soil_temperature_min 0.089647403
#> 15              soil_clay 0.077514274
#> 16             topo_slope 0.046389046
#> 17              soil_sand 0.043249031
#> 18         topo_elevation 0.029362052
#> 19  soil_temperature_mean 0.023888671
#> 20              soil_silt 0.007505508

#using it in variable selection with collinear()
selected.predictors <- cor_select(
  df = vi,
  response = "vi_mean", #don't forget the response!
  predictors = vi_predictors,
  preference_order = preference.order,
  max_cor = 0.75
  )

selected.predictors
#>  [1] "koppen_zone"            "soil_temperature_max"   "soil_temperature_range"
#>  [4] "swi_range"              "soil_soc"               "topo_diversity"        
#>  [7] "soil_clay"              "topo_slope"             "soil_sand"             
#> [10] "topo_elevation"         "soil_temperature_mean" 

#check their correlations
selected.predictors.cor <- cor_df(
  df = vi,
  response = "vi_mean",
  predictors = selected.predictors
)

#all correlations below max_cor
selected.predictors.cor
#> # A tibble: 55 × 3
#>    x                      y                      correlation
#>    <chr>                  <chr>                        <dbl>
#>  1 soil_temperature_max   koppen_zone                 -0.72 
#>  2 soil_temperature_range koppen_zone                 -0.698
#>  3 swi_range              koppen_zone                  0.677
#>  4 soil_sand              soil_clay                   -0.675
#>  5 soil_temperature_mean  soil_temperature_max         0.586
#>  6 soil_soc               soil_temperature_max        -0.578
#>  7 topo_slope             topo_diversity               0.538
#>  8 swi_range              soil_temperature_range      -0.5  
#>  9 soil_temperature_range soil_temperature_max         0.481
#> 10 soil_temperature_mean  soil_soc                    -0.431
#> # ℹ 45 more rows

#USING A CUSTOM FUNCTION
#custom function to compute RMSE between a predictor and a response
#x is a predictor name
#y is a response name
#df is a data frame with multiple predictors and one response
#must return a single number, with higher number indicating higher preference
#notice we use "one minus RMSE" to give higher rank to variables with lower RMSE
f_rmse <- function(x, y, df){

  xy <- df[, c(x, y)] |>
    na.omit() |>
    scale()

  1 - sqrt(mean((xy[, 1] - xy[, 2])^2))

}

preference.order <- preference_order(
  df = vi,
  response = "vi_mean",
  predictors = vi_predictors,
  f = f_rmse
)

preference.order
#>                 predictor  preference
#> 1             koppen_zone  0.55763087
#> 2      koppen_description  0.53878289
#> 3                swi_mean  0.47522602
#> 4            koppen_group  0.44462415
#> 5               soil_type  0.37435469
#> 6                 swi_max  0.32253839
#> 7               swi_range  0.16007942
#> 8                 swi_min  0.01110232
#> 9                soil_soc -0.09470103
#> 10         topo_diversity -0.14669333
#> 11   soil_temperature_min -0.18312100
#> 12              soil_clay -0.20072025
#> 13             topo_slope -0.25206567
#> 14              soil_silt -0.35089030
#> 15  soil_temperature_mean -0.51881865
#> 16         topo_elevation -0.52982498
#> 17              soil_sand -0.55354823
#> 18 soil_temperature_range -0.86221416
#> 19   soil_temperature_max -0.88683592
#> 20                soil_ph -0.93183434

#disable parallelization
future::plan(future::sequential)