Target-encoding methods

Methods to apply target-encoding to individual categorical variables. The functions implemented are:

fe_target_encoding_mean(): Each group is identified by the mean of the response over the group cases. White noise can be added via the noise argument. Columns encoded with this function are identified by the suffix "__encoded_mean". If noise is used, then the amount of noise is also added to the suffix.

fe_target_encoding_rank(): Each group is identified by the rank of the mean of the response variable over the group cases. The group with the lower mean receives the rank 1. White noise can be added via the noise argument. Columns encoded with this function are identified by the suffix "__encoded_rank". If noise is used, then the amount of noise is also added to the suffix.

fe_target_encoding_rnorm(): Each case in a group receives a value coming from a normal distribution with the mean and the standard deviation of the response over the cases of the group. The argument sd.width multiplies the standard deviation to reduce the spread of the obtained values. Columns encoded with this function are identified by the suffix "__encoded_rnorm_sd.width_X", where X is the amount of sd.width used.

fe_target_encoding_loo(): The suffix "loo" stands for "leave-one-out". Each case in a group is encoded as the average of the response over the other cases of the group. olumns encoded with this function are identified by the suffix "__encoded_loo".

Usage

fe_target_encoding_mean(
  data,
  response.name,
  categorical.variable.name,
  noise = 0,
  seed = 1,
  replace = FALSE,
  verbose = TRUE
)

fe_target_encoding_rnorm(
  data,
  response.name,
  categorical.variable.name,
  sd.width = 0.1,
  seed = 1,
  replace = FALSE,
  verbose = TRUE
)

fe_target_encoding_rank(
  data,
  response.name,
  categorical.variable.name,
  noise = 0,
  seed = 1,
  replace = FALSE,
  verbose = TRUE
)

fe_target_encoding_loo(
  data,
  response.name,
  categorical.variable.name,
  replace = FALSE,
  verbose = TRUE
)

fe_target_encoding_noise(data, categorical.variable.name, noise = 0, seed = 1)

Arguments

data

(required; data frame, tibble, or sf) A training data frame. Default: NULL

response.name

(required; character string) Name of the response. Must be a column name of data. Default: NULL

categorical.variable.name

(required; character) Name of the categorical variable to encode.

noise

(optional; numeric) Numeric with noise values in the range 0-1. Default: 0.

seed

(optional; integer) Random seed to facilitate reproducibility. Default: 1

replace

(optional; logical) Advanced option that changes the behavior of the function. Use only if you really know exactly what you need. If TRUE, it replaces each categorical variable with its encoded version, and returns the input data frame with the replaced variables.

verbose

(optional; logical) If TRUE, messages and plots generated during the execution of the function are displayed. Default: TRUE

sd.width

(optional; numeric) Numeric with multiplicator of the standard deviation of each group in the categorical variable, in the range 0.01-1. Default: 0.1

Value

The input data frame with a target-encoded variable.

Examples

if(interactive()){

#target_encoding mean
data(
  ecoregions,
  ecoregions_continuous_response,
  ecoregions_all_predictors
)

#the dataframe ecoregions contains a categorical variable named primary_productivity
unique(ecoregions$primary_productivity)

#Group mean
#########################################################

#transforming primary_productivity
ecoregions <- fe_target_encoding_mean(
  data = ecoregions,
  response.name = ecoregions_continuous_response,
  categorical.variable.name = "primary_productivity"
)

#the encoded variable is named primary_productivity__encoded_mean

#plotting it against the response
ggplot2::ggplot(data = ecoregions) +
  ggplot2::aes(
    x = plant_richness,
    y = primary_productivity__encoded_mean,
    color = primary_productivity
  ) +
  ggplot2::geom_point()

#Group rank
#########################################################

#transforming primary_productivity
ecoregions <- fe_target_encoding_rank(
  data = ecoregions,
  response.name = ecoregions_continuous_response,
  categorical.variable.name = "primary_productivity"
)

#the encoded variable is named primary_productivity__encoded_rank

#plotting it against the response
ggplot2::ggplot(data = ecoregions) +
  ggplot2::aes(
    x = plant_richness,
    y = primary_productivity__encoded_rank,
    color = primary_productivity
  ) +
  ggplot2::geom_point()

#Leave-one-out
#########################################################
#transforming primary_productivity
ecoregions <- fe_target_encoding_loo(
  data = ecoregions,
  response.name = ecoregions_continuous_response,
  categorical.variable.name = "primary_productivity"
)

#the encoded variable is named primary_productivity__encoded_loo

#plotting it against the response
ggplot2::ggplot(data = ecoregions) +
  ggplot2::aes(
    x = plant_richness,
    y = primary_productivity__encoded_loo,
    color = primary_productivity
  ) +
  ggplot2::geom_point()

#rnorm
#########################################################

#transforming primary_productivity
ecoregions <- fe_target_encoding_rnorm(
  data = ecoregions,
  response.name = ecoregions_continuous_response,
  categorical.variable.name = "primary_productivity"
)

#the encoded variable is named primary_productivity__encoded_rnorm_sd.width_0.1

#plotting it against the response
ggplot2::ggplot(data = ecoregions) +
  ggplot2::aes(
    x = plant_richness,
    y = primary_productivity__encoded_rnorm_sd.width_0.1,
    color = primary_productivity
  ) +
  ggplot2::geom_point()

}