Validate Argument predictors

Internal function to validate the predictors argument. Requires the argument 'df' to be validated with validate_df().

Validates the 'predictors' argument to ensure it complies with the requirements of the package functions. It performs the following actions:

Stops if 'df' is NULL.

Stops if 'df' is not validated.

If 'predictors' is NULL, uses column names of 'df' as 'predictors' in the 'df' data frame.

Print a message if there are names in 'predictors' not in the column names of 'df', and returns only the ones in 'df'.

Stop if the number of numeric columns in 'predictors' is smaller than 'min_numerics'.

Print a message if there are zero-variance columns in 'predictors' and returns a new 'predictors' argument without them.

Tags the vector with the attribute validated = TRUE to let the package functions skip the data validation.

Usage

validate_predictors(
  df = NULL,
  response = NULL,
  predictors = NULL,
  decimals = 4
)

Arguments

df

(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.

response

(optional, character string) Name of a numeric (non-numerics are ignored) response variable in df, only required if there are categorical variables named in the argument predictors. Default: NULL.

predictors

(optional; character vector) Names of the variables to select from df. If omitted, all numeric columns in df are used instead. If argument response is not provided, non-numeric variables are ignored. Default: NULL

decimals

(required, integer) Number of decimal places for the zero variance test. Smaller numbers will increase the number of variables detected as near-zero variance. Recommended values will depend on the range of the numeric variables in 'df'. Default: 4

Value

character vector: predictor names

See also

Author

Blas M. Benito, PhD

Examples

data(
  vi,
  vi_predictors
  )

#validating example data frame
vi <- validate_df(
  df = vi
)

#validating example predictors
vi_predictors <- validate_predictors(
  df = vi,
  predictors = vi_predictors
)

#tagged as validated
attributes(vi_predictors)$validated
#> [1] TRUE