Toy dataframe with varying levels of multicollinearity

Dataframe with known relationship between responses and predictors useful to illustrate multicollinearity concepts.

Usage

data(toy)

Format

dataframe with 2000 rows and 5 columns.

Details

Columns:

y: response variable generated from a * 0.75 + b * 0.25 + noise.

a: most important predictor of y, uncorrelated with b.

b: second most important predictor of y, uncorrelated with a.

c: generated from a + noise.

d: generated from (a + b)/2 + noise.

These are variance inflation factors of the predictors in toy. variable vif b 4.062 d 6.804 c 13.263 a 16.161

See also