Variance Inflation Factor

Computes the Variance Inflation Factor of all variables in a training data frame.

The Variance Inflation Factor for a given variable y is computed as 1/(1-R2), where R2 is the multiple R-squared of a multiple regression model fitted using y as response and all the remaining variables of the input data set as predictors. The equation can be interpreted as "the rate of perfect model's R-squared to the unexplained variance of this model".

The possible range of VIF values is (1, Inf]. A VIF lower than 10 suggest that removing y from the data set would reduce overall multicollinearity.

This function computes the Variance Inflation Factor (VIF) in three steps:

Computes the correlation matrix between all pairs of predictors using stats::cor().

Applies solve() to obtain the precision matrix, which is the inverse of the covariance matrix.

Uses diag() to extract the diagonal of the precision matrix, which contains the variance of the prediction of each predictor from all other predictors.

Usage

vif_df(df = NULL, response = NULL, predictors = NULL, encoding_method = "mean")

Arguments

df

(required; data frame or tibble) A data frame with numeric and/or character predictors predictors, and optionally, a response variable. Default: NULL.

response

(recommended, character string) Name of a numeric response variable. Character response variables are ignored. Please, see 'Details' to better understand how providing this argument or not leads to different results when there are character variables in 'predictors'. Default: NULL.

predictors

(optional; character vector) character vector with predictor names in 'df'. If omitted, all columns of 'df' are used as predictors. Default:'NULL'

encoding_method

(optional; character string). Name of the target encoding method to convert character and factor predictors to numeric. One of "mean", "rank", "loo", "rnorm" (see target_encoding_lab() for further details). Default: "mean"

Value

Data frame with predictor names and VIF values

Examples

if(interactive()){

data(
  vi,
  vi_predictors
)

#reduce size of vi to speed-up example execution
vi <- vi[1:1000, ]

#without response
#only numeric predictors are returned
df <- vif_df(
  df = vi,
  predictors = vi_predictors
)

df

#with response
#categorical and numeric predictors are returned
df <- vif_df(
  df = vi,
  response = "vi_mean",
  predictors = vi_predictors
)

df

}